refactor(rib): split update-group actor helpers
