+++ crates/rib/src/manager/update_groups.rs
@@ -44,6 +44,10 @@ use rustbgpd_wire::{
 use rustc_hash::FxHashMap;
 use tracing::{debug, info, warn};
 
+mod policy_transition;
+mod staging;
+mod views;
+
 use super::helpers::{LOCAL_PEER, routes_equal, vpn_routes_equal};
 use super::{PolicyFilteredRouteKey, RibManager, RtcMembership};
 use crate::adj_rib_out::AdjRibOut;
@@ -2300,172 +2304,6 @@ fn assert_scoped_visit_bound(visits: usize, scope: usize, residue: usize) {
 }
 
 impl RibManager {
-    /// Snapshot route identities for the strict clean transition inventory.
-    /// Dirty state or private-family participation rejects the optimization
-    /// before a member is moved or an envelope is emitted.
-    pub(in crate::manager) fn begin_clean_policy_transition_inventory(
-        &self,
-        source: usize,
-        destination: usize,
-    ) -> Option<Vec<(Prefix, u32)>> {
-        let clean = |group: &GroupRibOut| {
-            // Per-client-best groups are excluded from the ADR-0105 narrow
-            // fast path outright — its clean predicate (zero
-            // policy-filtered routes on both sides) contradicts the
-            // mitigation's reason to exist. Re-inclusion is demand-gated
-            // (ADR-0126 Decision 8).
-            !group.per_client_best
-                && group.dirty_members.is_empty()
-                && group.tombstones.is_empty()
-                && group.vpn_tombstones.is_empty()
-                && group.table.vpn_len() == 0
-                && group.policy_filtered.is_empty()
-                && group.vpn_policy_denied.is_empty()
-                && group.otc_blocked.is_empty()
-        };
-        let old = self.group_ribs.get(&source)?;
-        let new = self.group_ribs.get(&destination)?;
-        if !clean(old) || !clean(new) || old.table.len() != new.table.len() {
-            return None;
-        }
-        Some(
-            new.table
-                .iter()
-                .map(|route| (route.prefix, route.path_id))
-                .collect(),
-        )
-    }
-
-    /// Accumulate one bounded key chunk. A withdrawal, source flip, or table
-    /// drift rejects the complete optimized plan before emission — and so
-    /// does a control-form community (RFC 7947 §2.3.2) for any of the
-    /// cohort's RS ASNs (`rs_asns`), on EITHER side of policy: the source
-    /// residue drives per-target suppression/prepend even when policy
-    /// strips the tag from the post-policy route, and a policy-added tag
-    /// needs the per-target scrub. Untagged (the overwhelming case) is
-    /// proven in this same single walk, so rs-control members ride the
-    /// shared cohort at zero extra passes; a tagged inventory hands the
-    /// whole cohort to the authoritative per-peer path, whose emit seams
-    /// apply the per-target filter.
-    pub(in crate::manager) fn extend_clean_policy_transition_inventory(
-        &self,
-        source: usize,
-        destination: usize,
-        keys: &[(Prefix, u32)],
-        rs_asns: &[u32],
-        inventory: &mut CleanPolicyTransitionInventoryBuilder,
-    ) -> Option<()> {
-        use super::distribution::rs_control::rs_control_route_tagged;
-        let old = self.group_ribs.get(&source)?;
-        let new = self.group_ribs.get(&destination)?;
-        // Fold permit counts per chunk keyed by borrowed labels, then merge
-        // into the owned builder maps once: the per-route path used to clone
-        // the `Option<String>` label twice per route, which dominated this
-        // walk at reload-stall scale. The merge clones one label per
-        // distinct (label) / (source, label) pair per chunk instead.
-        let mut chunk_totals: FxHashMap<Option<&str>, u64> = FxHashMap::default();
-        let mut chunk_by_source: FxHashMap<IpAddr, FxHashMap<Option<&str>, u64>> =
-            FxHashMap::default();
-        // Run-length fold for the permit counters: tables interleave far
-        // fewer (source, label) flips than routes (contiguous prefix blocks
-        // from one source are the common shape), so accumulate runs and
-        // touch the maps once per flip instead of twice per route.
-        let mut run: Option<(IpAddr, Option<&str>, u64)> = None;
-        for &(prefix, path_id) in keys {
-            let route = new.table.get(&prefix, path_id)?;
-            let key = (prefix, path_id);
-            let prior = old.table.get(&prefix, path_id)?;
-            if prior.peer != route.peer {
-                return None;
-            }
-            let next_hop = new.nh_override(key);
-            if !routes_equal(prior, route) || old.nh_override(key) != next_hop {
-                if !rs_asns.is_empty() {
-                    let (old_communities, old_large) = old.source_control(key);
-                    let (new_communities, new_large) = new.source_control(key);
-                    let tagged = rs_asns.iter().any(|&rs_asn| {
-                        rs_control_route_tagged(
-                            route.communities(),
-                            route.large_communities(),
-                            rs_asn,
-                        ) || rs_control_route_tagged(old_communities, old_large, rs_asn)
-                            || rs_control_route_tagged(new_communities, new_large, rs_asn)
-                    });
-                    if tagged {
-                        return None;
-                    }
-                }
-                inventory.announce.push(route.clone());
-                inventory.next_hop_override.push(next_hop);
-            }
-
-            let label = new
-                .staged_labels
-                .get(&key)
-                .and_then(|label| label.as_deref());
-            run = Some(match run {
-                Some((peer, run_label, count)) if peer == route.peer && run_label == label => {
-                    (peer, run_label, count + 1)
-                }
-                Some((peer, run_label, count)) => {
-                    *chunk_totals.entry(run_label).or_default() += count;
-                    *chunk_by_source
-                        .entry(peer)
-                        .or_default()
-                        .entry(run_label)
-                        .or_default() += count;
-                    (route.peer, label, 1)
-                }
-                None => (route.peer, label, 1),
-            });
-        }
-        if let Some((peer, run_label, count)) = run {
-            *chunk_totals.entry(run_label).or_default() += count;
-            *chunk_by_source
-                .entry(peer)
-                .or_default()
-                .entry(run_label)
-                .or_default() += count;
-        }
-        for (label, count) in chunk_totals {
-            *inventory
-                .permit_totals
-                .entry(label.map(str::to_owned))
-                .or_default() += count;
-        }
-        for (peer, counts) in chunk_by_source {
-            let by_source = inventory.permit_by_source.entry(peer).or_default();
-            for (label, count) in counts {
-                *by_source.entry(label.map(str::to_owned)).or_default() += count;
-            }
-        }
-        Some(())
-    }
-
-    /// Apply the inventory's pre-aggregated permit counts to one member after
-    /// its writer slot has been reserved. This preserves the existing grouped
-    /// split-horizon counter semantics without another full-table walk.
-    pub(in crate::manager) fn apply_clean_policy_transition_counters(
-        &mut self,
-        peer: IpAddr,
-        inventory: &CleanPolicyTransitionInventory,
-    ) {
-        let own = inventory.permit_by_source.get(&peer);
-        let rows = inventory
-            .permit_totals
-            .iter()
-            .filter_map(|(label, total)| {
-                let count = total.saturating_sub(
-                    own.and_then(|counts| counts.get(label))
-                        .copied()
-                        .unwrap_or(0),
-                );
-                (count > 0).then_some((label.clone(), PolicyAction::Permit, count))
-            })
-            .collect::<Vec<_>>();
-        self.bump_export_counters(peer, &rows);
-    }
-
     /// The group id a registered peer is a member of, if any.
     pub(in crate::manager) fn grouped_member_of(&self, peer: IpAddr) -> Option<usize> {
         match self.update_groups.members.get(&peer) {
@@ -2596,679 +2434,6 @@ impl RibManager {
         self.rtc_vpn_filter(peer, self.peer_sendable_families.get(&peer))
     }
 
-    /// Run the shared staging pass for every live group over the pass's
-    /// changed prefixes. Deltas are committed to the group tables here;
-    /// the per-peer loop emits them per member via the source-flip
-    /// matrix. `memo` is the pass-scoped export memo shared with the
-    /// ungrouped fallback staging.
-    ///
-    /// A plain group stages exactly `best_changed`: its input is
-    /// Loc-RIB-best-or-nothing, so the narrow set is complete. A
-    /// per-client-best group's winner walk reads the candidate list,
-    /// not the Loc-RIB best — a candidate change can flip the winner
-    /// or the lane while the best stands — so it stages the widened
-    /// `best_changed ∪ all_affected` set, mirroring the ungrouped
-    /// per-client-best enumeration. The union is built once, and only
-    /// when such a group exists; winner-equality and lane suppression
-    /// make the widened pass a no-op wherever neither slot moved.
-    pub(in crate::manager) fn stage_update_groups(
-        &mut self,
-        best_changed: &HashSet<Prefix>,
-        all_affected: &HashSet<Prefix>,
-        memo: &mut super::distribution::ExportMemo,
-    ) -> HashMap<usize, GroupStageOutput> {
-        let mut staged = HashMap::new();
-        if self.group_ribs.is_empty() || (best_changed.is_empty() && all_affected.is_empty()) {
-            return staged;
-        }
-        let widened: Option<HashSet<Prefix>> = (!all_affected.is_empty()
-            && self.group_ribs.values().any(|group| group.per_client_best))
-        .then(|| best_changed.union(all_affected).copied().collect());
-        let gids: Vec<usize> = self.group_ribs.keys().copied().collect();
-        for gid in gids {
-            let prefixes = match (&widened, self.group_ribs.get(&gid)) {
-                (Some(widened), Some(group)) if group.per_client_best => widened,
-                _ => best_changed,
-            };
-            if prefixes.is_empty() {
-                continue;
-            }
-            let mut out = self.stage_group_prefixes(gid, prefixes, memo);
-            // Built here (the fanout path) and not inside the staging
-            // pass: `join_group`'s table-build pass discards its output.
-            out.build_shared_emit();
-            staged.insert(gid, out);
-        }
-        // The staging commit is the one lane mutation site outside the
-        // membership lifecycle (which refreshes via the gauge sweep).
-        self.refresh_lane_gauge();
-        staged
-    }
-
-    /// One shared export-tail pass for `gid` over `prefixes`, reusing
-    /// `distribute_single_best_prefix` with split horizon lifted out
-    /// (`ExportTarget::Group`) and the group table as the diff baseline
-    /// — the SAME body as the per-peer path, parameterized, never
-    /// copied (design risk 1). Deltas are committed before returning;
-    /// tombstones extend when a member is already dirty.
-    #[expect(
-        clippy::too_many_lines,
-        reason = "one staging pass keeps the plain single-best arm and the ADR-0126 \
-                  per-client-best arm over the same commit block"
-    )]
-    fn stage_group_prefixes(
-        &mut self,
-        gid: usize,
-        prefixes: &HashSet<Prefix>,
-        memo: &mut super::distribution::ExportMemo,
-    ) -> GroupStageOutput {
-        let mut out = GroupStageOutput::default();
-        let mut labeled_filtered: Vec<(PolicyFilteredRouteKey, Option<PolicyLabel>)> = Vec::new();
-        let mut lane_updates: Vec<(Prefix, Option<RunnerUp>)> = Vec::new();
-        {
-            let Some(group) = self.group_ribs.get(&gid) else {
-                return out;
-            };
-            // `share()`, not `clone()`: evaluations through the group
-            // handle must land in the installed chain's ADR-0096 term
-            // hit counters, exactly like the per-peer path's handle.
-            let chain = group.export_chain.as_ref().map(PolicyChain::share);
-            let mut announce: Vec<Route> = Vec::new();
-            let mut withdraw: Vec<(Prefix, u32)> = Vec::new();
-            let mut nh_flags: Vec<Option<NextHopAction>> = Vec::new();
-            let mut filtered: Vec<PolicyFilteredRouteKey> = Vec::new();
-            for prefix in prefixes {
-                let old_source = group.table.get(prefix, 0).map(|r| r.peer);
-                if group.per_client_best {
-                    // ADR-0126 Decision 2: first-permitted winner walk
-                    // + runner-up lane.
-                    let deltas_before = out.deltas.len();
-                    let stage = Self::distribute_group_per_client_best_prefix(
-                        &self.ribs,
-                        &self.unicast_prefix_peers,
-                        &group.table,
-                        &self.peer_is_rr_client,
-                        prefix,
-                        group.is_ebgp,
-                        group.interpret_rfc1997,
-                        group.is_rr_client,
-                        group.local_role,
-                        self.cluster_id,
-                        Some(&group.sendable),
-                        Some(&group.llgr),
-                        chain.as_ref(),
-                        memo,
-                        &mut out.evals,
-                        &mut announce,
-                        &mut withdraw,
-                        &mut nh_flags,
-                        &mut labeled_filtered,
-                        &mut out.otc_blocked,
-                    );
-                    // The winner announce (at most one — the walk
-                    // stages a single `path_id 0` winner), captured
-                    // for the lane-transition supersession decision.
-                    let winner_announce = announce.first().map(|route| route.peer);
-                    for (route, nh) in announce.drain(..).zip(nh_flags.drain(..)) {
-                        out.deltas.push(GroupDelta {
-                            prefix: *prefix,
-                            path_id: route.path_id,
-                            new: Some((route, nh)),
-                            old_source,
-                            policy_label: stage.winner_label.clone(),
-                            source_attrs: stage.winner_source_attrs.clone(),
-                            lane: stage.runner_up.clone(),
-                        });
-                    }
-                    for (p, path_id) in withdraw.drain(..) {
-                        out.deltas.push(GroupDelta {
-                            prefix: p,
-                            path_id,
-                            new: None,
-                            old_source,
-                            policy_label: None,
-                            source_attrs: None,
-                            lane: None,
-                        });
-                    }
-                    // Winner-side tag-only transition, mirroring the
-                    // plain arm's hook below: the pass staged nothing
-                    // for this prefix (winner equality-suppressed)
-                    // while the WINNER's source control communities
-                    // moved — rs-control members' verdicts on the
-                    // staged entry may flip with no wire change.
-                    if out.deltas.len() == deltas_before
-                        && let Some(transition) =
-                            group.rs_tag_transition(*prefix, stage.winner_source_attrs.as_ref())
-                    {
-                        out.rs_transitions.push(transition);
-                    }
-                    // Lane transition (ADR-0126 Decision 5), equality-
-                    // suppressed against the PRIOR lane entry:
-                    // `routes_equal` includes the source peer, so a
-                    // content-equal same-source reinstall is suppressed
-                    // while a source flip never is. The SOURCE control
-                    // communities must also be equal: lane entries
-                    // carry source attributes so rs-control tag
-                    // transitions extend to them — a control-community
-                    // change the chain erases post-policy would
-                    // otherwise flip `source(w)`'s suppress/prepend
-                    // verdict with no recorded transition. The
-                    // recomputed lane commits below regardless
-                    // (Decision 6).
-                    let prior = group.runner_up.get(prefix);
-                    let (content_unchanged, control_unchanged) = match (prior, &stage.runner_up) {
-                        (Some(old), Some(new)) => (
-                            routes_equal(&old.route, &new.route),
-                            source_control_input(old.source_attrs.as_ref())
-                                == source_control_input(new.source_attrs.as_ref()),
-                        ),
-                        (None, None) => (true, true),
-                        _ => (false, true),
-                    };
-                    if !(content_unchanged && control_unchanged) {
-                        // The member the transition emits toward: the
-                        // new entry's winner source, or the REPLACED
-                        // entry's for a retire (how the retire arm
-                        // knows `source(w)` — lane-only when the
-                        // winner is unchanged, the OLD winner source
-                        // on all-candidates-gone). Cleared when the
-                        // winner announce delta's own arms already
-                        // rewrite the target's slot: toward its NEW
-                        // source (which reads `GroupDelta::lane`
-                        // directly), and toward any OTHER member as
-                        // the flip-away `w'` announce.
-                        let target = stage
-                            .runner_up
-                            .as_ref()
-                            .map(|entry| entry.winner_source)
-                            .or_else(|| prior.map(|entry| entry.winner_source));
-                        let emit_target = target.filter(|target| match winner_announce {
-                            Some(source) => {
-                                if stage.runner_up.is_some() {
-                                    // Announce lane arm: superseded
-                                    // unless the winner source is
-                                    // unchanged (same-source content
-                                    // change — the winner arm skips
-                                    // its own source's slot).
-                                    old_source == Some(*target)
-                                } else {
-                                    // Retire arm: superseded when the
-                                    // winner flipped away from the
-                                    // target (it receives `w'`).
-                                    source == *target
-                                }
-                            }
-                            None => true,
-                        });
-                        out.lane_deltas.push(LaneDelta {
-                            prefix: *prefix,
-                            new: stage.runner_up.clone(),
-                            old_source: prior.map(|entry| entry.route.peer),
-                            emit_target,
-                            prior_source_attrs: prior.and_then(|entry| entry.source_attrs.clone()),
-                            content_unchanged,
-                        });
-                    }
-                    lane_updates.push((*prefix, stage.runner_up));
-                    continue;
-                }
-                // RFC 7947 decisions at the member-emit seams are made
-                // on the pre-policy SOURCE (the Loc-RIB best this pass
-                // stages from); capture its attributes for the deltas
-                // and the table residue.
-                let source_attrs = self.loc_rib.get(prefix).and_then(capture_source_attrs);
-                let deltas_before = out.deltas.len();
-                let mut target = super::distribution::ExportTarget::Group {
-                    evals: &mut out.evals,
-                    local_role: group.local_role,
-                    otc_blocked: &mut out.otc_blocked,
-                };
-                Self::distribute_single_best_prefix(
-                    &self.loc_rib,
-                    &group.table,
-                    &self.peer_is_rr_client,
-                    prefix,
-                    &mut target,
-                    group.is_ebgp,
-                    group.interpret_rfc1997,
-                    // Group staging is rs-control-agnostic: control
-                    // communities diverge per TARGET, so they are
-                    // enforced at the member-emit seams (LAN-474 —
-                    // matrix walk, resync, join/refresh replay), never
-                    // against the shared staged winner.
-                    None,
-                    group.is_rr_client,
-                    self.cluster_id,
-                    Some(&group.sendable),
-                    Some(&group.llgr),
-                    chain.as_ref(),
-                    None, // ORF disqualifies from grouping — never present here
-                    memo,
-                    &mut announce,
-                    &mut withdraw,
-                    &mut nh_flags,
-                    &mut filtered,
-                    false,
-                );
-                // Single-best stages at most one evaluation per prefix;
-                // its terminal-policy label tags the staged entry (or
-                // the denial residue) for join-time counter replay.
-                let label = out.evals.take_last().and_then(|(label, _, _)| label);
-                for (route, nh) in announce.drain(..).zip(nh_flags.drain(..)) {
-                    out.deltas.push(GroupDelta {
-                        prefix: *prefix,
-                        path_id: route.path_id,
-                        new: Some((route, nh)),
-                        old_source,
-                        policy_label: label.clone(),
-                        source_attrs: source_attrs.clone(),
-                        lane: None,
-                    });
-                }
-                for (p, path_id) in withdraw.drain(..) {
-                    out.deltas.push(GroupDelta {
-                        prefix: p,
-                        path_id,
-                        new: None,
-                        old_source,
-                        policy_label: None,
-                        source_attrs: None,
-                        lane: None,
-                    });
-                }
-                if out.deltas.len() == deltas_before
-                    && let Some(transition) =
-                        group.rs_tag_transition(*prefix, source_attrs.as_ref())
-                {
-                    out.rs_transitions.push(transition);
-                }
-                labeled_filtered.extend(filtered.drain(..).map(|key| (key, label.clone())));
-            }
-        }
-        let group = self
-            .group_ribs
-            .get_mut(&gid)
-            .expect("group staged above still exists");
-        for delta in &out.deltas {
-            group.apply_delta(delta);
-        }
-        // Lane commits live in this commit block ON PURPOSE: the
-        // `join_group` table-build pass discards the returned output
-        // but must still leave a fully populated lane behind.
-        for (prefix, entry) in lane_updates {
-            group.apply_lane(prefix, entry);
-        }
-        group.commit_rs_transitions(&out.rs_transitions);
-        group.record_otc_blocked(prefixes, &out.otc_blocked);
-        group.record_policy_filtered(prefixes, &labeled_filtered);
-        if !group.dirty_members.is_empty() {
-            let withdrawn: Vec<(Prefix, u32)> = out.withdrawn_keys().collect();
-            group.tombstones.extend(withdrawn);
-            // A member ALREADY dirty when a source flip stages onto it
-            // never reaches the per-member matrix (its pass takes the
-            // resync arm), so its member-scoped withdraw of the displaced
-            // route would be lost: the key stays IN the table (invisible
-            // to tombstones) and the resync announces table ∖ own-sourced.
-            // Record it as an extra (over-)withdraw at staging; the
-            // resync's `member_retains` guard drops it if the source
-            // flips back before the resync runs.
-            let dirty: Vec<IpAddr> = group.dirty_members.iter().copied().collect();
-            for member in dirty {
-                let lost: Vec<(Prefix, u32)> = out.member_scoped_withdraws(member).collect();
-                if !lost.is_empty() {
-                    self.pending_extra_withdraws
-                        .entry(member)
-                        .or_default()
-                        .unicast
-                        .extend(lost);
-                }
-            }
-            self.refresh_group_residue_gauge();
-        }
-        out
-    }
-
-    /// Run the shared VPN staging pass for every VPN-staging group over
-    /// the pass's changed RD+prefix identities. Deltas are committed to
-    /// the group tables here; `recompute_and_distribute_vpn` emits them
-    /// per member via the RT-pass source-flip matrix (Φ applied at emit
-    /// for RTC-negotiated groups).
-    pub(in crate::manager) fn stage_vpn_update_groups(
-        &mut self,
-        changed: &HashSet<VpnRouteKey>,
-    ) -> HashMap<usize, VpnGroupStageOutput> {
-        let mut staged = HashMap::new();
-        if changed.is_empty() || self.group_ribs.is_empty() {
-            return staged;
-        }
-        let gids: Vec<usize> = self
-            .group_ribs
-            .iter()
-            .filter(|(_, group)| group.stages_vpn())
-            .map(|(gid, _)| *gid)
-            .collect();
-        for gid in gids {
-            staged.insert(gid, self.stage_group_vpn_keys(gid, changed));
-        }
-        staged
-    }
-
-    /// One shared VPN export-tail pass for `gid` over `keys`, reusing
-    /// `stage_vpn_routes`'s single-best body with split horizon lifted
-    /// out (`ExportTarget::Group`) and the group table's VPN maps as the
-    /// diff baseline — the SAME body as the per-peer path, parameterized,
-    /// never copied (design risk 1). Deltas are committed before
-    /// returning; VPN tombstones extend when a member is already dirty.
-    fn stage_group_vpn_keys(
-        &mut self,
-        gid: usize,
-        keys: &HashSet<VpnRouteKey>,
-    ) -> VpnGroupStageOutput {
-        let mut out = VpnGroupStageOutput::default();
-        let mut denials: Vec<(VpnRouteKey, VpnDenialRecord)> = Vec::new();
-        {
-            let Some(group) = self.group_ribs.get(&gid) else {
-                return out;
-            };
-            // `share()`, not `clone()` — ADR-0096 term hit counters, as
-            // in `stage_group_prefixes`.
-            let chain = group.export_chain.as_ref().map(PolicyChain::share);
-            let mut announce: Vec<VpnRibRoute> = Vec::new();
-            let mut withdraw: Vec<VpnRibRouteKey> = Vec::new();
-            let mut ignored_otc_blocked = Vec::new();
-            // Reused single-key set: the staging body iterates a key set,
-            // but the delta needs per-key `old` capture and eval labels.
-            let mut key_set: HashSet<VpnRouteKey> = HashSet::with_capacity(1);
-            for key in keys {
-                key_set.clear();
-                key_set.insert(*key);
-                let mut target = super::distribution::ExportTarget::Group {
-                    evals: &mut out.evals,
-                    local_role: group.local_role,
-                    otc_blocked: &mut ignored_otc_blocked,
-                };
-                Self::stage_vpn_routes(
-                    &self.loc_rib,
-                    &self.ribs,
-                    &group.table,
-                    &self.peer_is_rr_client,
-                    &key_set,
-                    &mut target,
-                    group.is_ebgp,
-                    group.interpret_rfc1997,
-                    group.is_rr_client,
-                    self.cluster_id,
-                    Some(&group.sendable),
-                    Some(&group.llgr),
-                    // RT gate deferred to member emit: Φ is per-member
-                    // state, applied by the RT-pass matrix (the delta
-                    // carries `old` for exactly that).
-                    None,
-                    None, // ORR disqualifies from grouping
-                    0,    // Add-Path send disqualifies from grouping
-                    None, // Effective cap is inapplicable to grouped peers.
-                    &[],
-                    chain.as_ref(),
-                    &mut announce,
-                    &mut withdraw,
-                    false,
-                );
-                // Single-best stages at most one eval per key: a Permit
-                // labels the staged entry; a Deny lands in the persistent
-                // denial residue (join-time counter replay).
-                let last = out.evals.take_last();
-                let label = match &last {
-                    Some((label, PolicyAction::Permit, _)) => label.clone(),
-                    _ => None,
-                };
-                if let Some((label, PolicyAction::Deny, source)) = last {
-                    // The evaluated route is the Loc-RIB best (group
-                    // targets never carry ORR); its RTs let an RTC
-                    // member's join replay Φ-gate the denial.
-                    let rts = self
-                        .loc_rib
-                        .get_vpn(key)
-                        .map(|best| best.extended_communities().to_vec())
-                        .unwrap_or_default();
-                    denials.push((*key, (source, label, rts)));
-                }
-                if announce.is_empty() && withdraw.is_empty() {
-                    continue;
-                }
-                // Prior staged entry, cloned before commit — one clone
-                // per delta, total (slice 2's Φ(old) input).
-                let old = group
-                    .table
-                    .get_vpn(&VpnRibRouteKey {
-                        nlri_key: *key,
-                        path_id: 0,
-                    })
-                    .cloned();
-                for route in announce.drain(..) {
-                    out.deltas.push(VpnGroupDelta {
-                        key: *key,
-                        new: Some(route),
-                        old: old.clone(),
-                        policy_label: label.clone(),
-                    });
-                }
-                for rib_key in withdraw.drain(..) {
-                    debug_assert_eq!(rib_key.path_id, 0, "group table stages path 0 only");
-                    out.deltas.push(VpnGroupDelta {
-                        key: *key,
-                        new: None,
-                        old: old.clone(),
-                        policy_label: None,
-                    });
-                }
-            }
-        }
-        let group = self
-            .group_ribs
-            .get_mut(&gid)
-            .expect("group staged above still exists");
-        for delta in &out.deltas {
-            group.apply_vpn_delta(delta);
-        }
-        // Denial-residue transition scope: this pass's keys replace their
-        // prior records (the `record_policy_filtered` shape).
-        group.vpn_policy_denied.retain(|key, _| !keys.contains(key));
-        group.vpn_policy_denied.extend(denials);
-        if !group.dirty_members.is_empty() {
-            group
-                .vpn_tombstones
-                .extend(out.deltas.iter().filter(|d| d.new.is_none()).map(|d| d.key));
-            self.refresh_group_residue_gauge();
-        }
-        out
-    }
-
-    /// Unicast portion of a grouped member's resync update, assembled
-    /// from the group table with NO policy re-evaluation:
-    ///
-    /// - plain dirty: announce the member's derived view — `adv(m)`
-    ///   per staged key ([`GroupRibOut::adv_entry`], ADR-0126
-    ///   Decision 4: the staged winner for a non-source member, the
-    ///   lane runner-up substituting for the winner's own source —
-    ///   with the resolved entry's own nh/source-attr residue) —
-    ///   withdraw `tombstones ∖ still-retained` (over-withdraw is the
-    ///   safe direction — the member's missed sends are unknown);
-    /// - regroup (baseline present): one-shot diff — announce entries
-    ///   not `routes_equal` to the baseline, withdraw baseline keys the
-    ///   member no longer retains;
-    /// - force-only (RFC 8326 `GShut` refresh): re-announce everything,
-    ///   bypassing the equality diff, withdraw nothing.
-    ///
-    /// `rs_control` is `(rs_asn, member_asn)` for an
-    /// `rs_control_communities` member (LAN-474): table entries whose
-    /// captured SOURCE communities ([`GroupRibOut::source_control`])
-    /// suppress them toward this member are skipped — and withdrawn
-    /// when the member may have them on the wire (plain dirty, or a
-    /// baseline that records the key) — and announced entries are
-    /// rewritten (prepend from the source, scrub post-policy) per
-    /// target. Baselines snapshot through the same filter
-    /// ([`GroupRibOut::member_view_snapshot`]), so the regroup one-shot
-    /// diff compares wire state to wire state.
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "the resync assembly takes the member's full pending-withdraw context"
-    )]
-    pub(in crate::manager) fn assemble_group_resync(
-        group: &GroupRibOut,
-        member: IpAddr,
-        rs_control: Option<(u32, u32)>,
-        is_dirty: bool,
-        is_force: bool,
-        baseline: Option<&FxHashMap<(Prefix, u32), Route>>,
-        extras: Option<&HashSet<(Prefix, u32)>>,
-        announce: &mut Vec<Route>,
-        withdraw: &mut Vec<(Prefix, u32)>,
-        nh_override_flags: &mut Vec<Option<NextHopAction>>,
-    ) {
-        use super::distribution::rs_control::{rs_control_route_rewrite, rs_control_suppressed};
-        let mut suppressed_withdraws: HashSet<(Prefix, u32)> = HashSet::new();
-        for staged in group.table.iter() {
-            let key = (staged.prefix, staged.path_id);
-            // adv(m) resolves the slot ([`GroupRibOut::adv_entry`]):
-            // the staged winner for a non-source member, the lane
-            // runner-up for the winner's own source, nothing when the
-            // lane is empty — the substitution is never restated here
-            // (ADR-0126 Decision 6). The rs-control decision and the
-            // nh residue come from the RESOLVED entry: a substituted
-            // slot reads the lane's captured source attributes, not
-            // the staged winner's.
-            let Some(entry) = group.adv_entry(member, &staged.prefix, staged.path_id) else {
-                continue;
-            };
-            let (source_communities, source_large_communities) =
-                source_control_input(entry.source_attrs);
-            if rs_control_suppressed(source_communities, source_large_communities, rs_control) {
-                // Not on this member's wire going forward. Withdraw when
-                // it may be there now: always on a plain dirty resync
-                // (missed sends are unknown — over-withdraw is the safe
-                // direction), on a regroup diff only when the baseline
-                // proves the key was sent. Force-only re-announces and
-                // never withdraws.
-                if !is_force && (is_dirty || baseline.is_some_and(|base| base.contains_key(&key))) {
-                    suppressed_withdraws.insert(key);
-                }
-                continue;
-            }
-            let mut route = entry.route.clone();
-            rs_control_route_rewrite(&mut route, source_large_communities, rs_control);
-            if !is_force
-                && let Some(base) = baseline
-                && base.get(&key).is_some_and(|old| routes_equal(old, &route))
-            {
-                continue;
-            }
-            nh_override_flags.push(entry.nh.cloned());
-            announce.push(route);
-        }
-        // A key the member still retains — its derived view
-        // ([`GroupRibOut::adv_entry`]) holds a non-source-suppressed
-        // route there — must not be withdrawn; everything else in the
-        // candidate sets is a safe (possibly spurious) withdraw.
-        // Routing retention through `adv_entry` makes a lane
-        // substitution retain exactly like a staged entry: a recorded
-        // lane withdraw survives the filter while the lane is empty
-        // (it must be delivered) and drops once the lane refills (the
-        // substituted announce above replaces it — over-withdraw and
-        // over-announce compose safely, announce+withdraw of one key
-        // does not).
-        let member_retains = |key: &(Prefix, u32)| {
-            group.adv_entry(member, &key.0, key.1).is_some_and(|entry| {
-                let (communities, large_communities) = source_control_input(entry.source_attrs);
-                !rs_control_suppressed(communities, large_communities, rs_control)
-            })
-        };
-        let mut keys: HashSet<(Prefix, u32)> = suppressed_withdraws;
-        if let Some(base) = baseline {
-            keys.extend(base.keys().filter(|key| !member_retains(key)));
-        }
-        if is_dirty {
-            keys.extend(group.tombstones.iter().filter(|key| !member_retains(key)));
-        }
-        if let Some(extra) = extras {
-            keys.extend(extra.iter().filter(|key| !member_retains(key)));
-        }
-        withdraw.extend(keys);
-    }
-
-    /// VPN portion of a grouped member's resync update — the VPN sibling
-    /// of [`Self::assemble_group_resync`], same three shapes (plain
-    /// dirty / regroup one-shot diff / force), assembled from the group
-    /// table's VPN maps with NO policy re-evaluation, under the member's
-    /// CURRENT Φ (`filter`). Only called for groups that stage VPN.
-    ///
-    /// The Φ dimension (design §2.4): announce = table entries passing
-    /// `pass_m`, and retention is Φ-aware (a staged key failing Φ is
-    /// NOT retained). A membership delta missed while dirty is healed
-    /// through `extras`: the Φ-write seam records the keys leaving Φ as
-    /// extra (over-)withdraws — exact, instead of the design's blanket
-    /// failing-Φ withdraw term, which would put spurious withdraws on
-    /// clean regroup diffs too (spurious withdraws are RFC 4271 no-ops,
-    /// but the regroup one-shot diff is held to exact-stream parity).
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "the resync assembly takes the member's full pending-withdraw context"
-    )]
-    pub(in crate::manager) fn assemble_group_vpn_resync(
-        group: &GroupRibOut,
-        member: IpAddr,
-        filter: Option<&RtcMembership>,
-        is_dirty: bool,
-        is_force: bool,
-        baseline: Option<&FxHashMap<VpnRouteKey, VpnRibRoute>>,
-        extras: Option<&HashSet<VpnRouteKey>>,
-        vpn_announce: &mut Vec<VpnRibRoute>,
-        vpn_withdraw: &mut Vec<VpnRibRouteKey>,
-    ) {
-        for route in group.table.iter_vpn() {
-            if route.peer == member || !rt_passes(filter, route) {
-                continue;
-            }
-            if !is_force
-                && let Some(base) = baseline
-                && base
-                    .get(&route.nlri.key())
-                    .is_some_and(|old| vpn_routes_equal(old, route))
-            {
-                continue;
-            }
-            vpn_announce.push(route.clone());
-        }
-        let member_retains = |key: &VpnRouteKey| {
-            group
-                .table
-                .get_vpn(&VpnRibRouteKey {
-                    nlri_key: *key,
-                    path_id: 0,
-                })
-                .is_some_and(|route| route.peer != member && rt_passes(filter, route))
-        };
-        let mut keys: HashSet<VpnRouteKey> = HashSet::new();
-        if let Some(base) = baseline {
-            keys.extend(base.keys().filter(|key| !member_retains(key)));
-        }
-        if is_dirty {
-            keys.extend(
-                group
-                    .vpn_tombstones
-                    .iter()
-                    .filter(|key| !member_retains(key)),
-            );
-        }
-        if let Some(extra) = extras {
-            keys.extend(extra.iter().filter(|key| !member_retains(key)));
-        }
-        vpn_withdraw.extend(keys.into_iter().map(|key| VpnRibRouteKey {
-            nlri_key: key,
-            path_id: 0,
-        }));
-    }
-
     /// The membership-delta path for a grouped VPN member whose Φ
     /// changed (design §2.3, the RTC hard part): ONE walk of the group
     /// VPN table — old-Φ vs new-Φ per entry — emitting a member-scoped
@@ -3392,138 +2557,6 @@ impl RibManager {
         }
     }
 
-    /// Bump a member's export-policy counters by the group verdict:
-    /// `totals − own-sourced` per (policy, action), as integer adds —
-    /// identical totals to the per-peer path, which skips the eval for
-    /// routes the target itself sourced (split horizon precedes the
-    /// policy check).
-    pub(in crate::manager) fn apply_group_policy_counters(
-        &mut self,
-        peer: IpAddr,
-        evals: &GroupEvalAccumulator,
-    ) {
-        if evals.totals.is_empty() {
-            return;
-        }
-        let own = evals.per_source.get(&peer);
-        let rows: Vec<(Option<String>, PolicyAction, u64)> = evals
-            .totals
-            .iter()
-            .map(|(policy, action, total)| {
-                let own_n = own
-                    .and_then(|rows| {
-                        rows.iter()
-                            .find(|(p, a, _)| a == action && p == policy)
-                            .map(|(_, _, n)| *n)
-                    })
-                    .unwrap_or(0);
-                (
-                    policy.as_ref().map(ToString::to_string),
-                    *action,
-                    total.saturating_sub(own_n),
-                )
-            })
-            .collect();
-        self.bump_export_counters(peer, &rows);
-    }
-
-    /// Reconstruct a joining member's export counters from the group's
-    /// staged residue: one permit per `adv(m)` slot (labelled by its
-    /// retained terminal policy — the lane entry's where the member
-    /// sourced the winner of a per-client-best prefix), one deny per
-    /// persistent denial — own-sourced excluded on the denial side and
-    /// substituted-or-excluded on the permit side, exactly what the
-    /// per-peer initial-dump staging would have recorded, without
-    /// re-running policy. A route-refresh replay passes its `family` so
-    /// only the refreshed family's entries count (the per-peer path
-    /// re-evaluates only that family).
-    pub(in crate::manager) fn apply_group_join_counters(
-        &mut self,
-        peer: IpAddr,
-        gid: usize,
-        family: Option<(Afi, Safi)>,
-    ) {
-        // The member's Φ: the per-peer path's RT gate precedes the policy
-        // evaluation, so an RT-failed entry records NO eval — the replay
-        // must count only Φ-passing permits and denials (design §2.4).
-        let vpn_filter = self.rtc_vpn_filter(peer, self.peer_sendable_families.get(&peer));
-        let mut rows: Vec<(Option<String>, PolicyAction, u64)> = Vec::new();
-        {
-            let Some(group) = self.group_ribs.get(&gid) else {
-                return;
-            };
-            let mut bump = |policy: &Option<PolicyLabel>, action: PolicyAction| {
-                bump_counter_row(&mut rows, policy.as_ref(), action);
-            };
-            let in_family =
-                |prefix: &Prefix| family.is_none_or(|f| super::helpers::prefix_family(prefix) == f);
-            // One permit per `adv(m)` slot (`adv_entry`, ADR-0126
-            // Decision 4): the staged entry's label for a non-own
-            // slot; the LANE entry's label where the member sourced
-            // the winner — the runner-up's permit WAS evaluated
-            // (Decision 2's over-replay posture), and its retained
-            // terminal label is exactly what the walk recorded.
-            // Nothing for an own-sourced slot with no substitution.
-            for route in group.table.iter() {
-                if !in_family(&route.prefix) {
-                    continue;
-                }
-                let Some(adv) = group.adv_entry(peer, &route.prefix, route.path_id) else {
-                    continue;
-                };
-                bump(&adv.policy_label.cloned(), PolicyAction::Permit);
-            }
-            for (prefix, denials) in &group.policy_filtered {
-                if !in_family(prefix) {
-                    continue;
-                }
-                for (&(source_peer, _), label) in denials {
-                    if source_peer == peer {
-                        continue;
-                    }
-                    bump(label, PolicyAction::Deny);
-                }
-            }
-            // VPN dimension (only staged for non-RTC groups): permits
-            // from the staged labels, denies from the denial residue —
-            // own-sourced excluded on both sides (the per-peer path's
-            // split horizon returns before the policy evaluation).
-            let in_vpn_family = |key: &VpnRouteKey| {
-                family.is_none_or(|f| {
-                    let afi = match key.prefix.family() {
-                        VpnAddressFamily::V4 => Afi::Ipv4,
-                        VpnAddressFamily::V6 => Afi::Ipv6,
-                    };
-                    (afi, Safi::MplsVpn) == f
-                })
-            };
-            for route in group.table.iter_vpn() {
-                if route.peer == peer
-                    || !in_vpn_family(&route.nlri.key())
-                    || !rt_passes(vpn_filter.as_ref(), route)
-                {
-                    continue;
-                }
-                let label = group
-                    .vpn_staged_labels
-                    .get(&route.nlri.key())
-                    .cloned()
-                    .unwrap_or(None);
-                bump(&label, PolicyAction::Permit);
-            }
-            for (key, (source, label, rts)) in &group.vpn_policy_denied {
-                if *source == peer
-                    || !in_vpn_family(key)
-                    || !vpn_filter.as_ref().is_none_or(|m| m.matches_any(rts))
-                {
-                    continue;
-                }
-                bump(label, PolicyAction::Deny);
-            }
-        }
-        self.bump_export_counters(peer, &rows);
-    }
-
     /// Apply aggregated export-policy verdict rows to a peer's counters
     /// (integer adds — no per-(prefix × peer) prometheus lookups).
     fn bump_export_counters(&mut self, peer: IpAddr, rows: &[(Option<String>, PolicyAction, u64)]) {
@@ -3553,191 +2586,6 @@ impl RibManager {
                 *n,
             );
         }
-    }
-
-    /// Borrowed synthesized advertised-route view for a grouped peer
-    /// (`adv(m)` per staged key — the table entry or its lane
-    /// substitution, minus the slots the RFC 9234 backstop strips,
-    /// [`GroupRibOut::adv_entry_post_backstop`] — minus this member's
-    /// exact-export rejections, intersected with what its outbound
-    /// maxima admitted); `None` for ungrouped peers.
-    pub(in crate::manager) fn grouped_advertised_routes_iter(
-        &self,
-        peer: IpAddr,
-    ) -> Option<impl Iterator<Item = &Route>> {
-        let group = self.group_ribs.get(&self.grouped_member_of(peer)?)?;
-        let rejected = self.peer_unexportable.get(&peer);
-        let limits = self.outbound_prefix_limits.get(&peer);
-        Some(group.table.iter().filter_map(move |staged| {
-            // The substitution sits at the same (prefix, path_id 0)
-            // slot, so the member-local overlays key identically for
-            // the staged entry and its lane replacement.
-            let route = group
-                .adv_entry_post_backstop(peer, &staged.prefix, staged.path_id)?
-                .route;
-            (!rejected.is_some_and(|keys| {
-                keys.contains(&ExactExportKey::Unicast(route.prefix, route.path_id))
-            }) && limits.is_none_or(|limits| limits.admits_grouped(&route.prefix)))
-            .then_some(route)
-        }))
-    }
-
-    /// Ordered sibling used by resumable route listings. The group table's
-    /// persistent prefix index resumes at the cursor; member-local split
-    /// horizon and exact-export rejection remain streaming filters. A page
-    /// clones only yielded rows, but a high-exclusion member may inspect more
-    /// underlying group rows before filling that page.
-    pub(in crate::manager) fn grouped_advertised_routes_ordered_iter(
-        &self,
-        peer: IpAddr,
-        after: Option<RouteQueryKey>,
-    ) -> Option<impl Iterator<Item = &Route>> {
-        let group = self.group_ribs.get(&self.grouped_member_of(peer)?)?;
-        let rejected = self.peer_unexportable.get(&peer);
-        let limits = self.outbound_prefix_limits.get(&peer);
-        Some(
-            group
-                .table
-                .iter_ordered_from(after)
-                .filter_map(move |staged| {
-                    let route = group
-                        .adv_entry_post_backstop(peer, &staged.prefix, staged.path_id)?
-                        .route;
-                    // A lane substitution keeps the staged entry's prefix
-                    // position, so the persistent prefix-index ordering is
-                    // unchanged — but the cursor a caller resumes with
-                    // carries the YIELDED route's source peer, which
-                    // differs from the staged entry's at a substituted
-                    // slot. Re-filter on the yielded key so a page
-                    // boundary landing on a substitution resumes without
-                    // duplicating its prefix (the underlying iterator
-                    // filters on staged keys only).
-                    (after.is_none_or(|cursor| route_query_key(route) > cursor)
-                        && !rejected.is_some_and(|keys| {
-                            keys.contains(&ExactExportKey::Unicast(route.prefix, route.path_id))
-                        })
-                        && limits.is_none_or(|limits| limits.admits_grouped(&route.prefix)))
-                    .then_some(route)
-                }),
-        )
-    }
-
-    /// Materialized sibling of [`Self::grouped_advertised_routes_iter`] for
-    /// legacy full-snapshot and route-refresh callers.
-    pub(in crate::manager) fn grouped_advertised_routes(&self, peer: IpAddr) -> Option<Vec<Route>> {
-        self.grouped_advertised_routes_iter(peer)
-            .map(|routes| routes.cloned().collect())
-    }
-
-    /// Synthesized advertised-route count for a grouped peer; `None`
-    /// for ungrouped peers.
-    pub(in crate::manager) fn grouped_advertised_count(&self, peer: IpAddr) -> Option<usize> {
-        let group = self.group_ribs.get(&self.grouped_member_of(peer)?)?;
-        if self.outbound_prefix_limits.contains_key(&peer) {
-            // A limited member advertises its admitted projection, and that
-            // is decided per family — so reuse the single definition of
-            // usage rather than deriving the same quantity a second way.
-            return Some(
-                super::outbound_prefix_limits::LIMITED_FAMILIES
-                    .into_iter()
-                    .map(|afi| self.outbound_family_usage(peer, afi, true))
-                    .sum(),
-            );
-        }
-        let rejected = self.peer_unexportable.get(&peer).map_or(0, |keys| {
-            keys.iter()
-                .filter(|key| match key {
-                    // A rejected key reduces the count only while the
-                    // member's derived view holds a route at that slot
-                    // — staged or lane-substituted — that the RFC 9234
-                    // backstop delivers: a slot the backstop strips was
-                    // already subtracted by the synthesis's OTC term,
-                    // so subtracting its rejection too would count the
-                    // one missing slot twice.
-                    ExactExportKey::Unicast(prefix, path_id) => group
-                        .adv_entry_post_backstop(peer, prefix, *path_id)
-                        .is_some(),
-                    _ => false,
-                })
-                .count()
-        });
-        Some(group.advertised_count_for(peer).saturating_sub(rejected))
-    }
-
-    /// Synthesized VPN count for a grouped member after applying its sparse
-    /// exact-export rejection overlay. The group table and RTC membership
-    /// counters remain shared truths; only this member-local projection is
-    /// reduced.
-    pub(in crate::manager) fn grouped_vpn_advertised_count(&self, peer: IpAddr) -> Option<usize> {
-        let gid = self.vpn_grouped_member_of(peer)?;
-        let group = self.group_ribs.get(&gid)?;
-        let filter = self.rtc_vpn_filter(peer, self.peer_sendable_families.get(&peer));
-        let rejected = self.peer_unexportable.get(&peer).map_or(0, |keys| {
-            keys.iter()
-                .filter(|key| match key {
-                    ExactExportKey::Vpn(key) => group.table.get_vpn(key).is_some_and(|route| {
-                        route.peer != peer && rt_passes(filter.as_ref(), route)
-                    }),
-                    _ => false,
-                })
-                .count()
-        });
-        Some(
-            group
-                .vpn_advertised_count_for(peer)
-                .saturating_sub(rejected),
-        )
-    }
-
-    /// Per-family grouped advertised counts after subtracting the member's
-    /// sparse exact-export rejection overlay. Used by BMP stat 17 and the
-    /// public count query so neither reports the shared group table as wire
-    /// truth for a classic-message peer that rejected an oversized route.
-    pub(in crate::manager) fn grouped_family_counts(
-        &self,
-        peer: IpAddr,
-    ) -> Option<Vec<((Afi, Safi), u64)>> {
-        let gid = self.grouped_member_of(peer)?;
-        let group = self.group_ribs.get(&gid)?;
-        let mut counts = group.family_counts_for(peer);
-        let filter = self.rtc_vpn_filter(peer, self.peer_sendable_families.get(&peer));
-        if let Some(rejected) = self.peer_unexportable.get(&peer) {
-            for key in rejected {
-                let family = match key {
-                    // Subtract only while the member's derived view
-                    // holds a backstop-delivered route at the rejected
-                    // slot — the lane substitution shares the staged
-                    // key's prefix, so the family is identical either
-                    // way, and a backstop-stripped slot was already
-                    // subtracted by the synthesis's OTC term (the
-                    // `grouped_advertised_count` double-subtraction
-                    // rule, per family).
-                    ExactExportKey::Unicast(prefix, path_id)
-                        if group
-                            .adv_entry_post_backstop(peer, prefix, *path_id)
-                            .is_some() =>
-                    {
-                        Some(super::helpers::prefix_family(prefix))
-                    }
-                    ExactExportKey::Vpn(key)
-                        if group.table.get_vpn(key).is_some_and(|route| {
-                            route.peer != peer && rt_passes(filter.as_ref(), route)
-                        }) =>
-                    {
-                        Some(key.afi_safi())
-                    }
-                    _ => None,
-                };
-                let Some(family) = family else {
-                    continue;
-                };
-                if let Some((_, count)) = counts.iter_mut().find(|(entry, _)| *entry == family) {
-                    *count = count.saturating_sub(1);
-                }
-            }
-        }
-        counts.retain(|(_, count)| *count != 0);
-        Some(counts)
     }
 
     /// Compute (or recompute) a registered peer's update-group
@@ -4119,425 +2967,6 @@ impl RibManager {
             self.mark_outbound_dirty(peer);
         }
         self.distribute_changes_after_advertised_page_advance(&HashSet::new(), &HashSet::new());
-    }
-
-    /// Whether a source/destination group pair qualifies for the shared
-    /// batched authoritative transition: unicast-only on both sides and
-    /// differing only in export-chain content (the clean transition's
-    /// narrow shape, checked on the interned keys).
-    pub(in crate::manager) fn batched_transition_keys_qualify(
-        &self,
-        source: usize,
-        destination: usize,
-    ) -> bool {
-        let (Some(source_key), Some(destination_key)) = (
-            self.update_groups.group_key(source),
-            self.update_groups.group_key(destination),
-        ) else {
-            return false;
-        };
-        source_key.is_unicast_only()
-            && destination_key.is_unicast_only()
-            && source_key.same_staging_profile_except_chain(destination_key)
-    }
-
-    /// The fingerprint itself: disqualifiers first (design §1), then
-    /// the group key from RIB-staging inputs.
-    pub(in crate::manager) fn compute_update_group_membership(
-        &mut self,
-        peer: IpAddr,
-    ) -> GroupMembership {
-        let chain = self.export_policy_for(peer).cloned();
-        self.compute_update_group_membership_for_policy(peer, chain.as_ref())
-    }
-
-    /// Classify a prospective policy without changing the peer's installed
-    /// policy or runtime membership. Registry interning is append-only and
-    /// observational; no group table or wire state is touched here.
-    fn compute_update_group_membership_for_policy(
-        &mut self,
-        peer: IpAddr,
-        chain: Option<&PolicyChain>,
-    ) -> GroupMembership {
-        // Differential-oracle hook: force every peer onto the per-peer
-        // fallback path so identical scenarios can be compared grouped
-        // vs ungrouped. Reuses the policy-peer-context reason label.
-        #[cfg(test)]
-        if self.test_force_ungrouped {
-            return GroupMembership::PolicyPeerContext;
-        }
-        // Slow-peer isolation (LAN-470): a transport-flagged slow peer
-        // stays on the per-peer path so its backlog cannot drag the
-        // shared staging pass. Checked before the classifier — it
-        // overrides an otherwise-groupable fingerprint.
-        if self.slow_isolated_peers.contains(&peer) {
-            return GroupMembership::SlowPeer;
-        }
-        // ORF-receive negotiated ⇒ ungrouped from the start (the RFC
-        // 5291 §6 gate never meets grouping). Read from the live
-        // session record — `peer_orf_pending` drains as gates lift, so
-        // it can't answer "was ORF negotiated" later in the session.
-        let orf_negotiated = self
-            .live_sessions
-            .get(&peer)
-            .and_then(|sessions| sessions.last())
-            .is_some_and(|record| !record.negotiated_orf_recv.is_empty());
-        let input = self.update_group_classifier_input(
-            peer,
-            chain,
-            orf_negotiated || self.peer_orf_filters.contains_key(&peer),
-            false,
-        );
-        let fingerprint = match classify_update_group(input) {
-            UpdateGroupClassification::PolicyPeerContext => {
-                return GroupMembership::PolicyPeerContext;
-            }
-            UpdateGroupClassification::AddPathSend => return GroupMembership::AddPathSend,
-            UpdateGroupClassification::PerClientBest => return GroupMembership::PerClientBest,
-            UpdateGroupClassification::OrrVantage => return GroupMembership::OrrVantage,
-            UpdateGroupClassification::OrfInstalled => return GroupMembership::OrfInstalled,
-            UpdateGroupClassification::Groupable(fingerprint) => fingerprint,
-        };
-
-        // Clone released before the &mut intern below; chains are small
-        // and this runs at config/session-lifecycle frequency only.
-        let chain_idx = chain.map(|chain| self.update_groups.intern_chain(chain));
-        let key = GroupKey {
-            chain: chain_idx,
-            target_is_ebgp: fingerprint.target_is_ebgp,
-            target_is_rr_client: fingerprint.target_is_rr_client,
-            target_local_role: fingerprint.target_local_role,
-            interpret_rfc1997: fingerprint.interpret_rfc1997,
-            sendable_ipv4_unicast: fingerprint.sendable_ipv4_unicast,
-            sendable_ipv6_unicast: fingerprint.sendable_ipv6_unicast,
-            sendable_vpnv4: fingerprint.sendable_vpnv4,
-            sendable_vpnv6: fingerprint.sendable_vpnv6,
-            rtc_negotiated: fingerprint.rtc_negotiated,
-            per_client_best: fingerprint.per_client_best,
-            llgr_families: fingerprint.llgr_families,
-        };
-        GroupMembership::Grouped(self.update_groups.group_for(key))
-    }
-
-    /// Preflight the prospective group id for a clean unicast-only policy
-    /// transition and prove that only the chain dimension changes.
-    pub(in crate::manager) fn clean_policy_transition_destination(
-        &mut self,
-        peer: IpAddr,
-        policy: Option<&PolicyChain>,
-    ) -> Option<(usize, usize)> {
-        let GroupMembership::Grouped(source) = self.update_groups.members.get(&peer)? else {
-            return None;
-        };
-        let source = *source;
-        // A per-client-best source group never takes the ADR-0105 fast
-        // path — nor its unfenced destination prestage, whose only
-        // preflight is this function. Checked on the source cell: the
-        // ADR-0126 `per_client_best` key bit keeps source and
-        // destination in agreement, so rejecting the source here also
-        // keeps every destination cell the prestage creates plain.
-        // Re-inclusion is demand-gated (ADR-0126 Decision 8).
-        if self
-            .group_ribs
-            .get(&source)
-            .is_some_and(|group| group.per_client_best)
-        {
-            return None;
-        }
-        let GroupMembership::Grouped(destination) =
-            self.compute_update_group_membership_for_policy(peer, policy)
-        else {
-            return None;
-        };
-        let source_key = self.update_groups.group_key(source)?;
-        let destination_key = self.update_groups.group_key(destination)?;
-        (source != destination
-            && source_key.is_unicast_only()
-            && destination_key.is_unicast_only()
-            && source_key.same_staging_profile_except_chain(destination_key))
-        .then_some((source, destination))
-    }
-
-    /// Create an unowned prospective destination group and return its prefix
-    /// staging snapshot. An existing destination is already maintained and
-    /// therefore needs no staging (`None`).
-    pub(in crate::manager) fn begin_policy_transition_group(
-        &mut self,
-        gid: usize,
-        peer: IpAddr,
-        export_policy: Option<&PolicyChain>,
-    ) -> PolicyTransitionGroupStart {
-        if self.group_ribs.contains_key(&gid) {
-            return PolicyTransitionGroupStart::Maintained;
-        }
-        let group = GroupRibOut::new(
-            export_policy.map(PolicyChain::share),
-            self.peer_is_ebgp.get(&peer).copied().unwrap_or(false),
-            self.peer_interpret_rfc1997.contains(&peer),
-            self.peer_is_rr_client.get(&peer).copied().unwrap_or(false),
-            self.peer_local_roles.get(&peer).copied().flatten(),
-            self.peer_sendable_families
-                .get(&peer)
-                .cloned()
-                .unwrap_or_default(),
-            self.peer_advertised_llgr_families
-                .get(&peer)
-                .cloned()
-                .unwrap_or_default(),
-            // ADR-0126 Decision 1: derived from the key bit. In
-            // practice always false here — per-client-best SOURCE
-            // groups never pass the fast-path preflight (Decision 8),
-            // and the key bit keeps source and destination in
-            // agreement.
-            self.update_groups
-                .group_key(gid)
-                .is_some_and(|key| key.per_client_best),
-            self.loc_rib.len(),
-        );
-        self.group_ribs.insert(gid, group);
-        PolicyTransitionGroupStart::Created(
-            self.loc_rib
-                .iter()
-                .map(|route| route.prefix)
-                .collect::<HashSet<_>>()
-                .into_iter()
-                .collect(),
-        )
-    }
-
-    /// Stage one bounded prefix chunk into an unowned destination group.
-    pub(in crate::manager) fn stage_policy_transition_group_chunk(
-        &mut self,
-        gid: usize,
-        prefixes: &[Prefix],
-        memo: &mut super::distribution::ExportMemo,
-    ) {
-        let prefixes = prefixes.iter().copied().collect::<HashSet<_>>();
-        let _ = self.stage_group_prefixes(gid, &prefixes, memo);
-    }
-
-    /// Remove a partially or fully staged, still-unowned destination before
-    /// the caller hands the transition back to the authoritative per-peer path.
-    pub(in crate::manager) fn discard_uncommitted_policy_transition_group(
-        &mut self,
-        gid: usize,
-    ) -> bool {
-        let removable = self
-            .group_ribs
-            .get(&gid)
-            .is_none_or(|group| group.members.is_empty());
-        if removable {
-            self.group_ribs.remove(&gid);
-        }
-        removable
-    }
-
-    /// Commit a preflighted clean transition after every writer slot and exact
-    /// snapshot has been validated. No regroup baseline is needed because the
-    /// shared transition diff is already the authoritative old-to-new wire
-    /// delta accepted by every target writer.
-    pub(in crate::manager) fn commit_clean_policy_transition_member(
-        &mut self,
-        peer: IpAddr,
-        source: usize,
-        destination: usize,
-    ) {
-        self.leave_group_without_gauge_refresh(source, peer);
-        self.update_groups
-            .members
-            .insert(peer, GroupMembership::Grouped(destination));
-        self.install_group_member(destination, peer);
-        self.metrics.record_update_group_regroup();
-    }
-
-    /// Publish cohort-wide gauges once after the synchronous membership commit.
-    /// Refreshing them per member is both unobservable (queries cannot
-    /// interleave in the commit section) and quadratic for large cohorts.
-    pub(in crate::manager) fn finish_clean_policy_transition_commit(&mut self) {
-        self.refresh_update_group_gauges();
-        self.refresh_group_residue_gauge();
-    }
-
-    /// Build the shared old→new inventory for one batched authoritative
-    /// cohort transition ([`crate::update::RibUpdate::ReplacePeerExportPoliciesAuthoritatively`]):
-    /// every batch member of `source` moves to the freshly staged
-    /// `destination` (same staging profile except the chain — validated
-    /// by the caller). Returns `None` when the delta carries an RFC 7947
-    /// control-form community for one of the cohort's rs-control ASNs on
-    /// either side of policy — per-target suppress/prepend/scrub cannot
-    /// ride a shared payload, so the caller degrades the cohort to the
-    /// ordinary per-member machinery (the clean transition's
-    /// tagged-inventory posture).
-    ///
-    /// The shared announce carries each destination entry whose wire
-    /// form differs from the source entry at the same key
-    /// (equality-suppressed — O(actual policy diff), a content-equal
-    /// restage emits nothing); the shared withdraw carries the keys the
-    /// destination no longer stages (over-withdraw safe toward members
-    /// whose wire never held them). Per-member divergence is exactly the
-    /// ADR-0126 Decision 4 emit-time exception set: split horizon rides
-    /// the envelope's `announce_source_exclusion`, and the lane
-    /// substitution toward each winner's source rides the member-scoped
-    /// supplements built here. Counter aggregates are folded once so the
-    /// per-member replay is O(labels), not O(table) — the same rows
-    /// [`RibManager::apply_group_join_counters`] would derive per member.
-    #[expect(
-        clippy::too_many_lines,
-        reason = "one walk keeps the shared delta, the lane supplements, the \
-                  OTC/rs-control degrade gates, and the counter fold aligned"
-    )]
-    pub(in crate::manager) fn batched_transition_inventory(
-        source: &GroupRibOut,
-        destination: &GroupRibOut,
-        rs_asns: &[u32],
-    ) -> Option<BatchedTransitionInventory> {
-        use super::distribution::rs_control::rs_control_route_tagged;
-        // RFC 9234 OTC residue on either side means blocked staged
-        // winners (or lane runner-ups) whose per-member diagnostics and
-        // withdraw conversions belong to the pre-commit backstop this
-        // shared emission does not run — the per-member path owns those.
-        // A role-bearing fleet with no OTC-attributed routes (the
-        // rendered route-server default) carries no residue and shares.
-        if !source.otc_blocked.is_empty() || !destination.otc_blocked.is_empty() {
-            return None;
-        }
-        let attrs_tagged = |attrs: (&[u32], &[LargeCommunity])| {
-            rs_asns
-                .iter()
-                .any(|&rs_asn| rs_control_route_tagged(attrs.0, attrs.1, rs_asn))
-        };
-        let route_tagged = |route: &Route| {
-            rs_asns.iter().any(|&rs_asn| {
-                rs_control_route_tagged(route.communities(), route.large_communities(), rs_asn)
-            })
-        };
-        // The pre-commit backstop strips an OTC-blocked route from every
-        // emission; an emitted delta containing one cannot ride the
-        // shared payload, so its cohort degrades to the per-member path
-        // (which runs the backstop at its own seams).
-        let blocked =
-            |route: &Route| super::distribution::otc_egress_blocked(route, destination.local_role);
-        let mut announce: Vec<Route> = Vec::new();
-        let mut next_hop_override: Vec<Option<NextHopAction>> = Vec::new();
-        let mut supplements: FxHashMap<IpAddr, BatchedMemberSupplement> = FxHashMap::default();
-        let mut counters = BatchedTransitionCounters::default();
-        for route in destination.table.iter() {
-            let key = (route.prefix, route.path_id);
-            let next_hop = destination.nh_override(key);
-            let prior = source.table.get(&route.prefix, route.path_id);
-            let changed = match prior {
-                None => true,
-                Some(prior_route) => {
-                    !routes_equal(prior_route, route) || source.nh_override(key) != next_hop
-                }
-            };
-            if changed {
-                if blocked(route)
-                    || (!rs_asns.is_empty()
-                        && (route_tagged(route)
-                            || attrs_tagged(destination.source_control(key))
-                            || attrs_tagged(source.source_control(key))))
-                {
-                    return None;
-                }
-                next_hop_override.push(next_hop);
-                announce.push(route.clone());
-            }
-            // Counter fold: one permit per staged entry, labelled by its
-            // retained terminal policy; per-source rows feed the
-            // Decision 4 `totals − own + lane` synthesis below.
-            let label = destination.staged_labels.get(&key).cloned().unwrap_or(None);
-            counters.record_permit(route.peer, label);
-            // Member-scoped correction for the slot's own source — the
-            // one member the shared exclusion silences at this key. Its
-            // wire held `adv_source(m)` (the source group's lane
-            // substitution when it also sourced the old winner, the old
-            // entry otherwise); its target is the destination lane's
-            // substitution or nothing (ADR-0126 Decision 4).
-            let lane_new = destination.runner_up.get(&route.prefix);
-            let prior_source = prior.map(|prior_route| prior_route.peer);
-            let lane_old = source.runner_up.get(&route.prefix);
-            if let Some(entry) = lane_new {
-                let suppressed = prior_source == Some(route.peer)
-                    && lane_old.is_some_and(|prev| {
-                        routes_equal(&prev.route, &entry.route) && prev.nh == entry.nh
-                    });
-                if !suppressed {
-                    if blocked(&entry.route)
-                        || (!rs_asns.is_empty()
-                            && (route_tagged(&entry.route)
-                                || attrs_tagged(source_control_input(entry.source_attrs.as_ref()))
-                                || lane_old.is_some_and(|prev| {
-                                    attrs_tagged(source_control_input(prev.source_attrs.as_ref()))
-                                })))
-                    {
-                        return None;
-                    }
-                    supplements
-                        .entry(route.peer)
-                        .or_default()
-                        .announce
-                        .push((entry.route.clone(), entry.nh.clone()));
-                }
-            } else {
-                // Over-withdraw-safe: emit whenever the source group
-                // staged the key and the member's wire could hold
-                // anything there (a displaced other-sourced entry — the
-                // source-flip arm — or a retired substitution). A no-op
-                // when the wire was already empty.
-                let had_wire =
-                    prior.is_some() && (prior_source != Some(route.peer) || lane_old.is_some());
-                if had_wire {
-                    supplements
-                        .entry(route.peer)
-                        .or_default()
-                        .withdraw
-                        .push(key);
-                }
-            }
-        }
-        let mut withdraw: Vec<(Prefix, u32)> = Vec::new();
-        for route in source.table.iter() {
-            if destination
-                .table
-                .get(&route.prefix, route.path_id)
-                .is_none()
-            {
-                withdraw.push((route.prefix, route.path_id));
-            }
-        }
-        for entry in destination.runner_up.values() {
-            counters.record_lane(entry.winner_source, entry.policy_label.clone());
-        }
-        for denials in destination.policy_filtered.values() {
-            for (&(source_peer, _), label) in denials {
-                counters.record_deny(source_peer, label.clone());
-            }
-        }
-        Some(BatchedTransitionInventory {
-            announce: announce.into(),
-            next_hop_override: next_hop_override.into(),
-            withdraw,
-            supplements,
-            counters,
-        })
-    }
-
-    /// Replay one cohort member's export counters from the batched
-    /// inventory's pre-aggregated rows — the exact rows
-    /// [`Self::apply_group_join_counters`] would fold from a full table
-    /// walk (`totals − own-sourced + lane substitutions` on the permit
-    /// side, `totals − own-sourced` on the denial side), at O(labels)
-    /// per member instead of O(table).
-    pub(in crate::manager) fn apply_batched_transition_counters(
-        &mut self,
-        peer: IpAddr,
-        inventory: &BatchedTransitionInventory,
-    ) {
-        let rows = inventory.counters.rows_for(peer);
-        if !rows.is_empty() {
-            self.bump_export_counters(peer, &rows);
-        }
     }
 
     /// Build the classifier input for one peer.
